From os.Stdout to taking a print func
It's clear that we cannot effectively test Print if all it does is based
on side effects on os.Stdout.

Instead, we explicitly declare a dependecy on something that can do the
underlying print, so that we can mock it for tests.

We mock it with a function that stores printed strings into a local
variable.