Take range as input, allowing multiple test cases
We want to write multiple tests that verify that Print works as it
should. It is not easy to do so if we cannot control what it should
print.

We could evolve the tests by increasingly adding things to what we want,
and then patching the implementation accordingly, but that's not
sustainable, and we clearly cannot keep doing it exaustively up to 100.

Instead, we set the pattern for writing Table-Driven Tests, and pass in
explicly arguments to control what is to be printed, though those
arguments need not to be used yet in order to get us to green.